% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_file_coordinate_conversion.R
\name{coordinate_conversion}
\alias{coordinate_conversion}
\title{Converts all files in file list to '+init=epsg:4326'}
\usage{
coordinate_conversion(file_list)
}
\arguments{
\item{file_list}{A character vector or name of shape files loaded into R that you want to convert.}
}
\value{
Creates a new spatial object in the global environment with set coordinates, ending in '_coords'.
}
\description{
The function will take a list of imported shape files, and set the coordinates for mapping 
as seen in the Introduction to Spatial Data in R pdf by Lovelace, et. al. Function is not currently
in use, but may be useful in further mapping development
\url(https://cran.r-project.org/doc/contrib/intro-spatial-rl.pdf)
}
