% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_dataframe_to_spatial.R
\name{join_df_to_spatial}
\alias{join_df_to_spatial}
\title{Joins a dataframe to a spatial datatype based on key'}
\usage{
join_df_to_spatial(dataframe, dataframe_key, spatial_layer, spatial_layer_key)
}
\arguments{
\item{dataframe}{A dataframe to be joined.}

\item{dataframe_key}{The column name in the dataframe to be used as a key to join with the spatial data.}

\item{spatial_layer}{The spatial object to be joined.}

\item{spatial_layer_key}{The column name in the spatial object's dataframe (accessed by spatial_layer@data)
that will be used as a key to join with the dataframe.}
}
\value{
Returns a spatial dataframe object with all columns in the dataframe joined to the spatial databy the key.
}
\description{
The function will join a dataframe to spatial data, given the column names of the keys to 
join by in the dataframe and spatial data. It is best practice to make sure the data in the dataframe
all match a value in the spatial data. If there is no match, that row will not be joined to the spatial
data.
}
