% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_shape_files.R
\name{load_shape_files}
\alias{load_shape_files}
\title{Attempts to imports all .shp files in a folder}
\usage{
load_shape_files(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder where all shape files are.}
}
\value{
Character vector of names of shape files that were found in the containing folder.
}
\description{
The package will create a vector of names of shape files found in a folder and try to loads these shape files as objects in the global enviroment.
}
