% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_functions.R
\name{points_map_subregion}
\alias{points_map_subregion}
\title{For a specified subsection, maps out all existing values of a variable in a spatialpointsdataframe, with an optionally included label}
\usage{
points_map_subregion(spatial_dataframe, mapping_variable, subregion_column_name,
  subregion_value, label_variable)
}
\arguments{
\item{spatial_dataframe}{The spatialpointsdataframe to be mapped}

\item{mapping_variable}{The column name of the data to be mapped}

\item{subregion_column_name}{The column name of the subregion to be mapped}

\item{subregion_value}{The value in the aforementioned column name that will be mapped}

\item{label_variable}{The column name (typically the english name column) (OPTIONAL)}
}
\value{
Returns a tmap object, which can be viewed as a map layer
}
\description{
This function will take a spatialpintsdataframe, the variable in the spatial dataframe to be mapped, then the column name and value of 
the subregion of the spatialpointsdataframe to be mapped, along with the variable label (optional), and create a map layer filling the 
points associated with the mapping variable on a color scale. This is intented currently for only numeric data.
}
